Size the per-player state axis to the seated count (shape as source of truth)

+++ packages/catan-engine/src/catan_engine/mechanics/turn.py
@@ -31,9 +31,7 @@ def _end_turn_avail(layout: BoardLayout, state: BoardState, params: None) -> Mas
 def _end_turn_apply(
     layout: BoardLayout, state: BoardState, params: None, available: Mask
 ) -> tuple[BoardState, ResultCode]:
-    nxt = (state.current_player.astype(jnp.int32) + 1) % state.n_players.astype(
-        jnp.int32
-    )
+    nxt = (state.current_player.astype(jnp.int32) + 1) % state.n_players
     cand = state._replace(
         dice_roll=jnp.uint8(0),
         has_rolled=jnp.uint8(0),
